Fix rainbow currency animation and shop display
The rainbow currency for the current true final currency in the extended set had an otherwise-smooth color animation with an abrupt jump; this has been corrected and the animation slowed down to account for the new smooth transition.

Furthermore, my currency notation commit added one digit of precision onto all currency displays in order to enable currency displays in the modern currency format to be correct. This caused the final heroic item to have an expense value that spanned 2 lines. I've changed the minimum width of the effect column to correct this.