Show the total amount of online time played on a save
There was an unused variable in the game data called "realtimeRun" which was incremented endlessly and (after the Dark Matter update) never reset, effectively making it a counter for the total online time played since the start of the game.

Display the variable in the stats as such and shift the "X days ago" to the end of the "date of game start" line.