Fix the width of tooltips
A recent change resulted in tooltips taking 25% of the width of the display, regardless of font size, causing overflow to the right at smaller font sizes. This has been changed; the tooltips now fit within the column and their width is dependent on the font size, which helps to avoid overflow.