Deduplicate the "updateRequiredRows" function
There were two functions with exactly the same code, one of which named "updateRequiredRows" and the other named "renderRequirementsForCategoryType". The latter function was only called one unique time while the former is called from multiple different places.

Consolidate the two functions into one to improve maintainability for that part of the code.