Deduplicate code for heroic required tooltips
There are two identical parts of the code in two different functions, responsible for piecing together the correct tooltip for jobs and skills yet to become Heroic while that mechanic is unlocked.

This commit deduplicates and moves the code to its own function, which improves maintainability by not having to apply changes to the code in multiple different places.