Clarify the challenge 5 description post-Metaverse
After Metaverse, lifespan is infinite except when in challenges 5 or 6 (which ordinally reduces lifespan); however that is not accounted for in the description.

Add text to the challenge 5 description stating the lifespan buff from Metaverse is disabled when in that challenge. This text only shows after the first Metaverse reset.