Fix hypercube sacrifice to challenge altar
Sacrificing hypercubes to stop challenges being reset on collapse deducted the cost for the next level of essence multiplier.

If a player had enough hypercubes to stop challenges being reset, but not enough for the next level of essence multiplier, their hypercube count would become negative when sacrificing to prevent challenges being reset.

This is now fixed; the challenge altar sacrifice now always deducts its fixed sacrifice cost of 1e10, instead of the cost for the next level of essence multiplier.